+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 Changelog
 =========
 
+dev-master
+----------
+
+* **2014-12-07**: Added event dispatcher and URI_CONTEXT_BUILT event
+
 1.0
 ---
 
 * **2014-08-19**: Added validation for yaml routing auto configuration keys
 * **2014-08-10**: Replaced all instances of the term URL with URI
 * Separated this package from the RoutingAutoBundle
+